로그아웃 개발, 배포서버 로직 분리 및 수정
