+++ .env.production
@@ -1,3 +1,4 @@
 #실제 운영되는 서버, 로컬 https설정 제외
 VITE_APP_MODE=prod
 VITE_APP_BASE_URL=https://api.woowabirdieside.com
+VITE_APP_DOMAIN_URL=https://woowabirdieside.com
