+++ .env.stage
@@ -1,3 +1,4 @@
 ##실제 운영되는 서버, 로컬에 https설정이 되어 실행됨
 VITE_APP_MODE=stage
 VITE_APP_BASE_URL=https://api.woowabirdieside.com
+VITE_APP_DOMAIN_URL=https://woowabirdieside.com
