+++ src/constants/Constants.js
@@ -9,5 +9,6 @@ export const DEFAULT_SUCCESS_MSG = '완료되었습니다.';
 // export const BASE_URL = import.meta.env.VITE_APP_BASE_URL; // 로컬 - http://localhost:8080
 // export const API_BASE_URL = `${BASE_URL}/api`; // 로컬 - http://localhost:8080/api
 
+export const DOMAIN_URL = import.meta.env.VITE_APP_DOMAIN_URL;
 export const BASE_URL = import.meta.env.VITE_APP_BASE_URL; // 로컬 - http://localhost:8080
 export const API_BASE_URL = `${BASE_URL}`; // 로컬 - http://localhost:8080/api
