+++ src/token.js
@@ -1,3 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 export const TOKEN_DEV =
-  'eyJhbGciOiJIUzI1NiJ9.eyJjYXRlZ29yeSI6Il9fU2VjdXJlLWFjY2VzcyIsInVzZXJJZCI6NjcsInJvbGUiOiJST0xFX1VTRVIiLCJpYXQiOjE3MTM1ODM5NjAsImV4cCI6MTcxMzU4NTc2MH0.2L27QZ2obfD9GGiwqlaF_1vsoAlI1c424vc8QzXdF1Q';
+  'eyJhbGciOiJIUzI1NiJ9.eyJjYXRlZ29yeSI6Il9fU2VjdXJlLWFjY2VzcyIsInVzZXJJZCI6MTE4LCJyb2xlIjoiUk9MRV9VU0VSIiwiaWF0IjoxNzEzNjE3ODk5LCJleHAiOjE3MTM2MTk2OTl9.0idNNYhqPRtRHBmIYuipjeL5-4EVTcMz1ORzKR-kZKQ';
+
+export const getTokenDev = () => localStorage.getItem('TOKEN_DEV');
+export const setTokenDev = (token) => localStorage.setItem('TOKEN_DEV', token);
+export const clearTokenDev = () => localStorage.removeItem('TOKEN_DEV');
